requirements.txt
loguru
gunicorn
google-genai
python-multipart
numpy<2.0
accelerate==0.30.1
fastapi 
uvicorn 
pydantic